Move Dashboard composables in corresponding files
